image_generator/requirements.txt
google-genai
Pillow
google-cloud-aiplatform 
google-cloud-storage
xhtml2pdf
markdown2